requirements.txt
streamlit>=1.30.0
tensorflow-cpu>=2.15.0
Pillow==12.1.1
numpy==1.26.4
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0